refactor `Solver` class into a function
It felt awkward to have to instantiate a class to run
the solver. This makes calling the solver a simple function
call. Additionally, the board passed to the solver is
no longer modified. A new solved board is returned from
the function.

The solver function now returns `None` if the given
board is unsolvable.